Fix integer overflow in av1_iadst4
Unscaled intermediate values in 4-point ADST transforms
(av1_iadst4 and av1_fadst4) could exceed INT32_MAX on
extreme inputs, causing a signed integer overflow.

This commit promotes the intermediate state variables to
int64_t to safely handle the accumulation before the
final downscale. A regression test is also included.

Bug: 471847850

+++ av1/common/av1_inv_txfm1d.c
@@ -657,12 +657,12 @@ void av1_iadst4(const int32_t *input, int32_t *output, int8_t cos_bit,
                 const int8_t *stage_range) {
   int bit = cos_bit;
   const int32_t *sinpi = sinpi_arr(bit);
-  int32_t s0, s1, s2, s3, s4, s5, s6, s7;
+  int64_t s0, s1, s2, s3, s4, s5, s6, s7;
 
-  int32_t x0 = input[0];
-  int32_t x1 = input[1];
-  int32_t x2 = input[2];
-  int32_t x3 = input[3];
+  int64_t x0 = input[0];
+  int64_t x1 = input[1];
+  int64_t x2 = input[2];
+  int64_t x3 = input[3];
 
   if (!(x0 | x1 | x2 | x3)) {
     output[0] = output[1] = output[2] = output[3] = 0;
